Move project page images to projects folder
