+++ src/pages/projects/jupiter.tsx
@@ -1,13 +1,13 @@
 import Navbar from '@/components/Navbar';
 import Footer from '@/components/Footer';
 import Head from 'next/head';
-import Rectangle from '@/../public/images/rectangle.png';
-import Cover from '@/../public/jupiter/cover.png';
 import Image from 'next/image';
+import Gradient from '@/../public/images/gradient.png';
 
-import search from '@/../public/jupiter/icons8-search-100.png';
-import calendar from '@/../public/jupiter/icons8-calendar-100.png';
-import community from '@/../public/jupiter/icons8-community-100.png';
+import Cover from '@/../public/projects/jupiter/cover.png';
+import search from '@/../public/projects/jupiter/icons8-search-100.png';
+import calendar from '@/../public/projects/jupiter/icons8-calendar-100.png';
+import community from '@/../public/projects/jupiter/icons8-community-100.png';
 
 const features = [
   {
@@ -42,7 +42,7 @@ const Jupiter = () => (
     </Head>
     <Navbar royal={true} className="relative z-20" />
     <div className="mx-8 lg:mx-16 xl:mx-32 pt-6 px-6 flex flex-col items-center gap-8 rounded-3xl text-white relative overflow-hidden">
-      <Image src={Rectangle} alt="project background" fill className="-z-20" />
+      <Image src={Gradient} alt="project background" fill className="-z-20" />
       <p className="px-16 py-2 rounded-full border-2 border-white whitespace-nowrap font-kallisto drop-shadow">
         JUPITER
       </p>
