+++ src/pages/projects/trends.tsx
@@ -1,13 +1,13 @@
 import Navbar from '@/components/Navbar';
 import Footer from '@/components/Footer';
 import Head from 'next/head';
-import Rectangle from '@/../public/images/rectangle.png';
-import Cover from '@/../public/trends/cover.png';
 import Image from 'next/image';
+import Gradient from '@/../public/images/gradient.png';
 
-import input from '@/../public/trends/icons8-input-100.png';
-import aggregate from '@/../public/trends/icons8-aggregate-100.png';
-import scale from '@/../public/trends/icons8-scale-100.png';
+import Cover from '@/../public/projects/trends/cover.png';
+import input from '@/../public/projects/trends/icons8-input-100.png';
+import aggregate from '@/../public/projects/trends/icons8-aggregate-100.png';
+import scale from '@/../public/projects/trends/icons8-scale-100.png';
 
 const features = [
   {
@@ -41,7 +41,7 @@ const Trends = () => (
     </Head>
     <Navbar royal={true} className="relative z-20" />
     <div className="mx-8 lg:mx-16 xl:mx-32 pt-6 px-6 flex flex-col items-center gap-8 rounded-3xl text-white relative overflow-hidden">
-      <Image src={Rectangle} alt="project background" fill className="-z-20" />
+      <Image src={Gradient} alt="project background" fill className="-z-20" />
       <p className="px-16 py-2 rounded-full border-2 border-white whitespace-nowrap font-kallisto drop-shadow">
         TRENDS
       </p>
